requirements.txt
Flask
google-generativeai
reportlab
